Add packages to keep relevant files together
