--- src/main/java/AddException.java
@@ -1,5 +1,0 @@
-public class AddException extends DenzException {
-    public AddException(String message) {
-        super(message);
-    }
-}
