--- src/main/java/ByeCommand.java
@@ -1,10 +1,0 @@
-public class ByeCommand extends Command{
-    @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) {
-        ui.showBye();
-    }
-    @Override
-    public boolean isExit() {
-        return true;
-    }
-}
