--- src/main/java/ByeException.java
@@ -1,5 +1,0 @@
-public class ByeException extends DenzException {
-    public ByeException(String message) {
-        super(message);
-    }
-}
