--- src/main/java/CorruptLineException.java
@@ -1,5 +1,0 @@
-public class CorruptLineException extends IllegalArgumentException {
-    public CorruptLineException(String message) {
-        super(message);
-    }
-}
