--- src/main/java/DeleteException.java
@@ -1,5 +1,0 @@
-public class DeleteException extends DenzException {
-    public DeleteException(String message) {
-        super(message);
-    }
-}
