--- src/main/java/DenzException.java
@@ -1,5 +1,0 @@
-public class DenzException extends Exception {
-    public DenzException(String message) {
-        super(message);
-    }
-}
