--- src/main/java/IndexException.java
@@ -1,5 +1,0 @@
-public class IndexException extends DenzException{
-    public IndexException(String message) {
-        super(message);
-    }
-}
