--- src/main/java/MarkException.java
@@ -1,5 +1,0 @@
-public class MarkException extends DenzException {
-    public MarkException(String message) {
-        super(message);
-    }
-}
