--- src/main/java/NoOpCommand.java
@@ -1,6 +1,0 @@
-public class NoOpCommand extends Command {
-    @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage) {
-        return;
-    }
-}
