--- src/main/java/Storage.java
@@ -1,61 +1,0 @@
-import java.io.IOException;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.nio.file.Files;
-import java.nio.file.StandardOpenOption;
-import java.util.ArrayList;
-import java.util.List;
-//For storing data into a File
-public class Storage {
-    private final Path path;
-    public Storage(String filePath){
-        this.path = Paths.get(filePath);
-    }
-    public TaskList load(){
-        ArrayList<Task> tasks = new ArrayList<>();
-        try {
-            //if parent directory not made yet, creates parent directory
-            if (path.getParent() != null) {
-                Files.createDirectories(path.getParent());
-            }
-            //if filename don't exist, create file and return empty list
-            if (!Files.exists(path)) {
-                Files.createFile(path);
-                return new TaskList(tasks);
-            }
-            List<String> lines = Files.readAllLines(path);
-            for (String line : lines) {
-                if (line.isBlank()) {
-                    continue;
-                }
-                try {
-                    //converts line into a task and adds to list
-                    tasks.add(TaskIO.fromSaveLine(line));
-                } catch (IllegalArgumentException e) {
-                    System.out.println("Skipping bad line " + line);
-                }
-            }
-        } catch (IOException e) {
-            System.out.println("Error reading file: " + e.getMessage());
-        }
-        return new TaskList(tasks);
-    }
-    public void save(TaskList tasks) {
-        List<String> lines = new ArrayList<>();
-        for (Task t : tasks.getList()) {
-            lines.add(TaskIO.toSaveLine(t));
-        }
-        try {
-            if (path.getParent() != null) {
-                Files.createDirectories(path.getParent());
-            }
-            //if file exist, remove all contents and rewrite into that file, else create new empty file and begin writing
-            Files.write(path, lines,
-                    StandardOpenOption.CREATE,
-                    StandardOpenOption.TRUNCATE_EXISTING);
-        } catch (IOException e) {
-            System.out.println("Error writing save file: " + e.getMessage());
-        }
-    }
-
-}
