+++ src/main/java/denz/exception/AddException.java
@@ -1,0 +1,7 @@
+package denz.exception;
+
+public class AddException extends DenzException {
+    public AddException(String message) {
+        super(message);
+    }
+}
