+++ src/main/java/denz/exception/ByeException.java
@@ -1,0 +1,7 @@
+package denz.exception;
+
+public class ByeException extends DenzException {
+    public ByeException(String message) {
+        super(message);
+    }
+}
