+++ src/main/java/denz/exception/CorruptLineException.java
@@ -1,0 +1,6 @@
+package denz.exception;
+public class CorruptLineException extends IllegalArgumentException {
+    public CorruptLineException(String message) {
+        super(message);
+    }
+}
