+++ src/main/java/denz/exception/DeleteException.java
@@ -1,0 +1,7 @@
+package denz.exception;
+
+public class DeleteException extends DenzException {
+    public DeleteException(String message) {
+        super(message);
+    }
+}
