+++ src/main/java/denz/exception/DenzException.java
@@ -1,0 +1,7 @@
+package denz.exception;
+
+public class DenzException extends Exception {
+    public DenzException(String message) {
+        super(message);
+    }
+}
