+++ src/main/java/denz/exception/IndexException.java
@@ -1,0 +1,7 @@
+package denz.exception;
+
+public class IndexException extends DenzException {
+    public IndexException(String message) {
+        super(message);
+    }
+}
