+++ src/main/java/denz/exception/MarkException.java
@@ -1,0 +1,7 @@
+package denz.exception;
+
+public class MarkException extends DenzException {
+    public MarkException(String message) {
+        super(message);
+    }
+}
