+++ src/main/java/denz/model/Todo.java
@@ -1,0 +1,13 @@
+package denz.model;
+
+public class Todo extends Task {
+    private String Description;
+    private boolean isDone;
+    public Todo(String Description) {
+        super(Description);
+    }
+    @Override
+    public String toString(){
+        return "[T] " + super.toString();
+    }
+}
